refactoring presenter package (made it viewmodel again), adding swipe-to-refresh layout
