+++ app/src/main/java/com/ringov/yamblzweather/viewmodel/about/AboutViewModel.java
@@ -1,0 +1,10 @@
+package com.ringov.yamblzweather.viewmodel.about;
+
+import com.ringov.yamblzweather.viewmodel.base.BaseViewModel;
+
+/**
+ * Created by ringov on 12.07.17.
+ */
+
+public class AboutViewModel extends BaseViewModel {
+}
